feat!(lsp): refactor lsp and treesitter configuration
