--- lua/config/lsp/servers/lua_ls.lua
@@ -1,72 +1,0 @@
-return {
-  servers = {
-    lua_ls = {
-      mason = false,
-      -- enabled = false,
-      -- cmd = { "/home/folke/projects/lua-language-server/bin/lua-language-server" },
-      single_file_support = true,
-      settings = {
-        Lua = {
-          workspace = {
-            checkThirdParty = false,
-          },
-          completion = {
-            workspaceWord = true,
-            callSnippet = "Both",
-          },
-          misc = {
-            parameters = {
-              -- "--log-level=trace",
-            },
-          },
-          hover = { expandAlias = false },
-          hint = {
-            enable = true,
-            setType = false,
-            paramType = true,
-            paramName = "Disable",
-            semicolon = "Disable",
-            arrayIndex = "Disable",
-          },
-          doc = {
-            privateName = { "^_" },
-          },
-          type = {
-            castNumberToInteger = true,
-          },
-          diagnostics = {
-            disable = { "incomplete-signature-doc", "trailing-space", "missing-fields" },
-            -- enable = false,
-            groupSeverity = {
-              strong = "Warning",
-              strict = "Warning",
-            },
-            groupFileStatus = {
-              ["ambiguity"] = "Opened",
-              ["await"] = "Opened",
-              ["codestyle"] = "None",
-              ["duplicate"] = "Opened",
-              ["global"] = "Opened",
-              ["luadoc"] = "Opened",
-              ["redefined"] = "Opened",
-              ["strict"] = "Opened",
-              ["strong"] = "Opened",
-              ["type-check"] = "Opened",
-              ["unbalanced"] = "Opened",
-              ["unused"] = "Opened",
-            },
-            unusedLocalExclude = { "_*" },
-          },
-          format = {
-            enable = true,
-            defaultConfig = {
-              indent_style = "space",
-              indent_size = "2",
-              continuation_indent_size = "2",
-            },
-          },
-        },
-      },
-    }
-  },
-}
