--- lua/config/lsp/servers/pyright.lua
@@ -1,18 +1,0 @@
-return {
-  servers = {
-    pyright = {
-      settings = {
-        python = {
-          analysis = {
-            typeCheckingMode = "basic",
-            diagnosticMode = "workspace",
-            inlayHints = {
-              variableTypes = true,
-              functionReturnTypes = true,
-            },
-          },
-        },
-      },
-    }
-  },
-}
