--- lua/plugins/lsp.lua
@@ -1,95 +1,0 @@
-return {
-  {
-    "folke/lazydev.nvim",
-    ft = "lua", -- only load on lua files
-    opts = {
-      library = {
-        -- See the configuration section for more details
-        -- Load luvit types when the `vim.uv` word is found
-        { path = "luvit-meta/library", words = { "vim%.uv" } },
-      },
-    },
-  },
-  { "Bilal2453/luvit-meta", lazy = true }, -- optional `vim.uv` typings
-
-  -- LSP Configuration
-
-  {
-    "williamboman/mason.nvim",
-    cmd = "Mason",
-    keys = { { "<leader>cm", "<cmd>Mason<cr>", desc = "Mason" } },
-    build = ":MasonUpdate",
-    opts_extend = { "ensure_installed" },
-    opts = {
-      ensure_installed = {
-        "shfmt",
-        "delve",
-        "js-debug-adapter",
-      },
-    },
-    ---@param opts MasonSettings | {ensure_installed: string[]}
-    config = function(_, opts)
-      require("mason").setup(opts)
-      local mr = require("mason-registry")
-      mr:on("package:install:success", function()
-        vim.defer_fn(function()
-          -- trigger FileType event to possibly load this newly installed LSP server
-          require("lazy.core.handler.event").trigger({
-            event = "FileType",
-            buf = vim.api.nvim_get_current_buf(),
-          })
-        end, 100)
-      end)
-
-      mr.refresh(function()
-        for _, tool in ipairs(opts.ensure_installed) do
-          local p = mr.get_package(tool)
-          if not p:is_installed() then p:install() end
-        end
-      end)
-    end,
-  },
-
-  {
-    "neovim/nvim-lspconfig",
-    branch = "master",
-    event = { "BufReadPre", "BufNewFile" },
-    dependencies = {
-      "williamboman/mason.nvim",
-      "williamboman/mason-lspconfig.nvim",
-      "someone-stole-my-name/yaml-companion.nvim",
-      "b0o/schemastore.nvim",
-    },
-    opts = require("config.lsp").opts(),
-    config = require("config.lsp").config,
-  },
-
-  {
-    "jay-babu/mason-nvim-dap.nvim",
-    config = function()
-      require("mason-nvim-dap").setup({
-        -- Get the list of installed servers from mason-lspconfig
-        ensure_installed = {
-          "js-debug-adapter", -- coding.config.dap.javascript
-          "delve",            -- coding.config.dap.go
-        },
-        automatic_installation = false,
-      })
-    end,
-  },
-
-  -- LSP Capabilities
-
-  "mfussenegger/nvim-jdtls",
-
-  "lvimuser/lsp-inlayhints.nvim",
-
-  {
-    "ray-x/lsp_signature.nvim",
-    opts = {
-      floating_window = false,               -- show hint in a floating window, set to false for virtual text only mode
-      floating_window_above_cur_line = true, -- try to place the floating above the current line when possible Note:
-      hint_scheme = "Comment",               -- highlight group for the virtual text
-    },
-  },
-}
